Replace NumeroSwitch by universal map control
